+++ pallets/ddc-clusters/src/lib.rs
@@ -81,6 +81,12 @@ pub mod pallet {
 		type StakerCreator: StakerCreator<Self, BalanceOf<Self>>;
 		type Currency: LockableCurrency<Self::AccountId, Moment = BlockNumberFor<Self>>;
 		type WeightInfo: WeightInfo;
+		#[pallet::constant]
+		type MinErasureCodingRequiredLimit: Get<u32>;
+		#[pallet::constant]
+		type MinErasureCodingTotalLimit: Get<u32>;
+		#[pallet::constant]
+		type MinReplicationTotalLimit: Get<u32>;
 	}
 
 	#[pallet::event]
@@ -111,6 +117,9 @@ pub mod pallet {
 		NodeAuthContractCallFailed,
 		NodeAuthContractDeployFailed,
 		NodeAuthNodeAuthorizationNotSuccessful,
+		ErasureCodingRequiredDidNotMeetMinimum,
+		ErasureCodingTotalNotMeetMinimum,
+		ReplicationTotalDidNotMeetMinimum
 	}
 
 	#[pallet::storage]
@@ -171,6 +180,9 @@ pub mod pallet {
 							.props
 							.node_provider_auth_contract
 							.clone(),
+						erasure_coding_required: 4,
+						erasure_coding_total: 6,
+						replication_total: 3
 					},
 					self.clusters_gov_params
 						.iter()
@@ -297,6 +309,9 @@ pub mod pallet {
 			let mut cluster =
 				Clusters::<T>::try_get(cluster_id).map_err(|_| Error::<T>::ClusterDoesNotExist)?;
 			ensure!(cluster.manager_id == caller_id, Error::<T>::OnlyClusterManager);
+			ensure!(cluster_params.erasure_coding_required >= T::MinErasureCodingRequiredLimit::get(), Error::<T>::ErasureCodingRequiredDidNotMeetMinimum);
+			ensure!(cluster_params.erasure_coding_total >= T::MinErasureCodingTotalLimit::get(), Error::<T>::ErasureCodingTotalNotMeetMinimum);
+			ensure!(cluster_params.replication_total >= T::MinReplicationTotalLimit::get(), Error::<T>::ReplicationTotalDidNotMeetMinimum);
 			cluster.set_params(cluster_params).map_err(Into::<Error<T>>::into)?;
 			Clusters::<T>::insert(cluster_id, cluster);
 			Self::deposit_event(Event::<T>::ClusterParamsSet { cluster_id });