+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [D] Changes is `Cere Dev` Runtime
 
 ## [VNext]
-
+- [C,D] `pallet-ddc-clusters`: Added Erasure coding and Replication in cluster params
+- 
 ## [5.3.0]
 
 ### Changed